+++ packages/core/src/manglers/html-attributes.ts
@@ -20,7 +20,10 @@ export type HtmlAttributeManglerOptions = {
   attrNamePattern?: string | string[];
 
   /**
-   * A list of HTML attribute names that should not be used.
+   * A list of strings and patterns of HTML attributes names that should not be
+   * used.
+   *
+   * Patterns are supported since v0.1.7.
    *
    * @default `[]`
    * @since v0.1.0