+++ packages/core/src/name-generator.class.ts
@@ -8,7 +8,7 @@ import type { Char, CharSet } from "./types";
  */
 export default class NameGenerator {
   /**
-   * The default characters set used by a {@link @NameGenerator}.
+   * The default set of characters used by {@link @NameGenerator}s.
    *
    * @since v0.1.0
    */
@@ -18,9 +18,9 @@ export default class NameGenerator {
   ];
 
   /**
-   * The list of reserved names.
+   * The list of reserved names and patterns.
    */
-  private readonly reserved: string[];
+  private readonly reserved: RegExp[];
 
   /**
    * The list of characters available to generate names with.
@@ -39,7 +39,20 @@ export default class NameGenerator {
    * Create a new {@link NameGenerator} that can be used to generate short,
    * safe, and unique strings.
    *
-   * @param reserved A list of reserved names.
+   * The `reserved` parameter can be used to specify strings that should not be
+   * generated. Each string is interpreted as a case sensitive Regular
+   * Expression that must match exactly. E.g. the reserved string "fa" will be
+   * transformed in the Regular Expression `/^fa$/` and hence only prevent the
+   * exact string "fa" from being generated. On the other hand, the reserved
+   * pattern "a.*", transformed into `/^a.*$/`, will prevent any string starting
+   * with an "a" (including just "a") from being generated.
+   *
+   * If you need to reserve a character that has a special meaning in Regular
+   * Expressions you need to escape it. E.g. the reserved string "a\\." will be
+   * transformed into `/^a\.$/` and will prevent the string "a." from being
+   * generated.
+   *
+   * @param reserved A list of reserved names or expressions.
    * @param charSet The character set to be used.
    * @since v0.1.0
    */
@@ -47,7 +60,7 @@ export default class NameGenerator {
     reserved: string[] = [],
     charSet: CharSet = NameGenerator.DEFAULT_CHARSET,
   ) {
-    this.reserved = reserved;
+    this.reserved = reserved.map((rawExpr) => new RegExp(`^${rawExpr}$`));
     this.charSet = charSet;
   }
 
@@ -59,11 +72,21 @@ export default class NameGenerator {
    */
   nextName(): string {
     this.current = this.tick(this.current);
-    if (this.reserved.includes(this.current)) {
+    if (this.isReserved(this.current)) {
       return this.nextName();
     } else {
       return this.current;
     }
+  }
+
+  /**
+   * Check with a string is reserved in the {@link NameGenerator}.
+   *
+   * @param s The string of interest.
+   * @returns `true` if `s` is reserved, `false` otherwise.
+   */
+  private isReserved(s: string): boolean {
+    return this.reserved.some((expr) => expr.test(s));
   }
 
   /**